# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * A professional REST API for user management, posts, and comments. Built with Node.js, Express, and PostgreSQL. Supports full CRUD operations, pagination, filtering, search, and structured error handling.

 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
  uptime: number;
  timestamp: string;
}

export type ErrorResponseDetailsItem = {
  field?: string;
  message?: string;
};

export interface ErrorResponse {
  error: string;
  message: string;
  details?: ErrorResponseDetailsItem[];
}

export interface MessageResponse {
  message: string;
}

export interface PaginationMeta {
  total: number;
  page: number;
  limit: number;
  totalPages: number;
  hasNextPage: boolean;
  hasPrevPage: boolean;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  user: "user",
  moderator: "moderator",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  bio?: string | null;
  avatarUrl?: string | null;
  isActive: boolean;
  postCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface UserResponse {
  data: User;
}

export interface UserListResponse {
  data: User[];
  meta: PaginationMeta;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  admin: "admin",
  user: "user",
  moderator: "moderator",
} as const;

export interface CreateUserBody {
  /**
   * @minLength 2
   * @maxLength 100
   */
  name: string;
  email: string;
  role?: CreateUserBodyRole;
  /** @maxLength 500 */
  bio?: string;
  avatarUrl?: string;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  admin: "admin",
  user: "user",
  moderator: "moderator",
} as const;

export interface UpdateUserBody {
  /**
   * @minLength 2
   * @maxLength 100
   */
  name?: string;
  email?: string;
  role?: UpdateUserBodyRole;
  /** @maxLength 500 */
  bio?: string;
  avatarUrl?: string;
  isActive?: boolean;
}

export type PostStatus = (typeof PostStatus)[keyof typeof PostStatus];

export const PostStatus = {
  draft: "draft",
  published: "published",
  archived: "archived",
} as const;

export interface Post {
  id: number;
  title: string;
  slug: string;
  content: string;
  excerpt?: string | null;
  status: PostStatus;
  category?: string | null;
  tags?: string[];
  viewCount: number;
  commentCount: number;
  authorId: number;
  author?: User | null;
  createdAt: string;
  updatedAt: string;
}

export interface PostResponse {
  data: Post;
}

export interface PostListResponse {
  data: Post[];
  meta: PaginationMeta;
}

export type CreatePostBodyStatus =
  (typeof CreatePostBodyStatus)[keyof typeof CreatePostBodyStatus];

export const CreatePostBodyStatus = {
  draft: "draft",
  published: "published",
  archived: "archived",
} as const;

export interface CreatePostBody {
  /**
   * @minLength 3
   * @maxLength 255
   */
  title: string;
  /** @minLength 10 */
  content: string;
  /** @maxLength 500 */
  excerpt?: string;
  status?: CreatePostBodyStatus;
  /** @maxLength 100 */
  category?: string;
  tags?: string[];
  authorId: number;
}

export type UpdatePostBodyStatus =
  (typeof UpdatePostBodyStatus)[keyof typeof UpdatePostBodyStatus];

export const UpdatePostBodyStatus = {
  draft: "draft",
  published: "published",
  archived: "archived",
} as const;

export interface UpdatePostBody {
  /**
   * @minLength 3
   * @maxLength 255
   */
  title?: string;
  /** @minLength 10 */
  content?: string;
  /** @maxLength 500 */
  excerpt?: string;
  status?: UpdatePostBodyStatus;
  /** @maxLength 100 */
  category?: string;
  tags?: string[];
}

export interface Comment {
  id: number;
  content: string;
  postId: number;
  authorId: number;
  author?: User | null;
  createdAt: string;
  updatedAt: string;
}

export interface CommentResponse {
  data: Comment;
}

export interface CommentListResponse {
  data: Comment[];
  meta: PaginationMeta;
}

export interface CreateCommentBody {
  /**
   * @minLength 1
   * @maxLength 1000
   */
  content: string;
  authorId: number;
}

export interface UpdateCommentBody {
  /**
   * @minLength 1
   * @maxLength 1000
   */
  content: string;
}

export type ListUsersParams = {
  page?: number;
  limit?: number;
  /**
   * Search by name or email
   */
  search?: string;
  role?: ListUsersRole;
};

export type ListUsersRole = (typeof ListUsersRole)[keyof typeof ListUsersRole];

export const ListUsersRole = {
  admin: "admin",
  user: "user",
  moderator: "moderator",
} as const;

export type GetUserPostsParams = {
  page?: number;
  limit?: number;
};

export type ListPostsParams = {
  page?: number;
  limit?: number;
  /**
   * Search by title or content
   */
  search?: string;
  status?: ListPostsStatus;
  category?: string;
  sortBy?: ListPostsSortBy;
  order?: ListPostsOrder;
};

export type ListPostsStatus =
  (typeof ListPostsStatus)[keyof typeof ListPostsStatus];

export const ListPostsStatus = {
  draft: "draft",
  published: "published",
  archived: "archived",
} as const;

export type ListPostsSortBy =
  (typeof ListPostsSortBy)[keyof typeof ListPostsSortBy];

export const ListPostsSortBy = {
  createdAt: "createdAt",
  updatedAt: "updatedAt",
  title: "title",
  viewCount: "viewCount",
} as const;

export type ListPostsOrder =
  (typeof ListPostsOrder)[keyof typeof ListPostsOrder];

export const ListPostsOrder = {
  asc: "asc",
  desc: "desc",
} as const;

export type GetPostCommentsParams = {
  page?: number;
  limit?: number;
};