# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: >
    A professional REST API for user management, posts, and comments.
    Built with Node.js, Express, and PostgreSQL. Supports full CRUD operations,
    pagination, filtering, search, and structured error handling.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List all posts
      description: Returns a paginated list of published posts with optional search, filter, and sort
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          schema:
            type: string
          description: Search by title or content
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, archived]
        - name: category
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, updatedAt, title, viewCount]
            default: createdAt
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Paginated list of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /posts/{id}:
    get:
      operationId: getPostById
      tags: [posts]
      summary: Get post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updatePost
      tags: [posts]
      summary: Update post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostBody"
      responses:
        "200":
          description: Post updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /posts/{id}/comments:
    get:
      operationId: getPostComments
      tags: [comments]
      summary: Get comments for a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Post comments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentListResponse"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentResponse"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /comments/{id}:
    put:
      operationId: updateComment
      tags: [comments]
      summary: Update a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCommentBody"
      responses:
        "200":
          description: Comment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentResponse"
        "404":
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comment deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
      required: [error, message]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        hasNextPage:
          type: boolean
        hasPrevPage:
          type: boolean
      required: [total, page, limit, totalPages, hasNextPage, hasPrevPage]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, user, moderator]
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        postCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, email, role, isActive, postCount, createdAt, updatedAt]
...
    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        content:
          type: string
        excerpt:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, published, archived]
        category:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        viewCount:
          type: integer
        commentCount:
          type: integer
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/User"
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, slug, content, status, viewCount, commentCount, authorId, createdAt, updatedAt]
    PostResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Post"
      required: [data]
    PostListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, meta]
    CreatePostBody:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 255
        content:
          type: string
          minLength: 10
        excerpt:
          type: string
          maxLength: 500
        status:
          type: string
          enum: [draft, published, archived]
          default: draft
        category:
          type: string
          maxLength: 100
        tags:
          type: array
          items:
            type: string
        authorId:
          type: integer
      required: [title, content, authorId]
    UpdatePostBody:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 255
        content:
          type: string
          minLength: 10
        excerpt:
          type: string
          maxLength: 500
        status:
          type: string
          enum: [draft, published, archived]
        category:
          type: string
          maxLength: 100
        tags:
          type: array
          items:
            type: string
    Comment:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        postId:
          type: integer
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/User"
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, content, postId, authorId, createdAt, updatedAt]
    CommentResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Comment"
      required: [data]
    CommentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, meta]
    CreateCommentBody:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000
        authorId:
          type: integer
      required: [content, authorId]
    UpdateCommentBody:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000
      required: [content]