# lib/api-zod/src/generated/types/createCommentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * A professional REST API for user management, posts, and comments. Built with Node.js, Express, and PostgreSQL. Supports full CRUD operations, pagination, filtering, search, and structured error handling.

 * OpenAPI spec version: 1.0.0
 */

export interface CreateCommentBody {
  /**
   * @minLength 1
   * @maxLength 1000
   */
  content: string;
  authorId: number;
}